requirements.txt
playwright==1.53.0
openai==1.59.7
pyyaml==6.0.3
click==8.2.1
pandas==3.0.0
